+++ src/ds323x/datetime.rs
@@ -19,11 +19,13 @@ where
         let mut data = [0; 8];
         self.iface.read_data(&mut data)?;
 
-        let year = year_from_registers(
-            data[Register::MONTH as usize + 1],
-            data[Register::YEAR as usize + 1],
-        );
-        let month = packed_bcd_to_decimal(data[Register::MONTH as usize + 1] & !BitFlags::CENTURY);
+        let century = data[Register::MONTH as usize + 1] & BitFlags::CENTURY;
+        if century != 0 {
+            return Err(Error::InvalidDeviceCentury);
+        }
+
+        let year = 2000 + (packed_bcd_to_decimal(data[Register::YEAR as usize + 1]) as u16);
+        let month = packed_bcd_to_decimal(data[Register::MONTH as usize + 1]);
         let day = packed_bcd_to_decimal(data[Register::DOM as usize + 1]);
         let hour = hours_from_register(data[Register::HOURS as usize + 1]);
         let minute = packed_bcd_to_decimal(data[Register::MINUTES as usize + 1]);
@@ -36,10 +38,9 @@ where
     }
 
     fn set_datetime(&mut self, datetime: &NaiveDateTime) -> Result<(), Self::Error> {
-        if datetime.year() < 2000 || datetime.year() > 2100 {
+        if !(2000..=2099).contains(&datetime.year()) {
             return Err(Error::InvalidInputData);
         }
-        let (month, year) = month_year_to_registers(datetime.month() as u8, datetime.year() as u16);
         let mut payload = [
             Register::SECONDS,
             decimal_to_packed_bcd(datetime.second() as u8),
@@ -47,8 +48,8 @@ where
             hours_to_register(Hours::H24(datetime.hour() as u8))?,
             datetime.weekday().number_from_sunday() as u8,
             decimal_to_packed_bcd(datetime.day() as u8),
-            month,
-            year,
+            decimal_to_packed_bcd(datetime.month() as u8),
+            decimal_to_packed_bcd((datetime.year() - 2000) as u8),
         ];
         self.iface.write_data(&mut payload)
     }
@@ -59,21 +60,50 @@ where
     DI: ReadData<Error = Error<E>> + WriteData<Error = Error<E>>,
 {
     fn seconds(&mut self) -> Result<u8, Self::Error> {
+        let data = self.iface.read_register(Register::MONTH)?;
+
+        let century = data & BitFlags::CENTURY;
+        if century != 0 {
+            return Err(Error::InvalidDeviceCentury);
+        }
+
         self.read_register_decimal(Register::SECONDS)
     }
 
     fn minutes(&mut self) -> Result<u8, Self::Error> {
+        let data = self.iface.read_register(Register::MONTH)?;
+
+        let century = data & BitFlags::CENTURY;
+        if century != 0 {
+            return Err(Error::InvalidDeviceCentury);
+        }
+
         self.read_register_decimal(Register::MINUTES)
     }
 
     fn hours(&mut self) -> Result<Hours, Self::Error> {
+        let data = self.iface.read_register(Register::MONTH)?;
+
+        let century = data & BitFlags::CENTURY;
+        if century != 0 {
+            return Err(Error::InvalidDeviceCentury);
+        }
+
         let data = self.iface.read_register(Register::HOURS)?;
         Ok(hours_from_register(data))
     }
 
     fn time(&mut self) -> Result<NaiveTime, Self::Error> {
+        let data = self.iface.read_register(Register::MONTH)?;
+
+        let century = data & BitFlags::CENTURY;
+        if century != 0 {
+            return Err(Error::InvalidDeviceCentury);
+        }
+
         let mut data = [0; 4];
         self.iface.read_data(&mut data)?;
+
         let hour = hours_from_register(data[Register::HOURS as usize + 1]);
         let minute = packed_bcd_to_decimal(data[Register::MINUTES as usize + 1]);
         let second = packed_bcd_to_decimal(data[Register::SECONDS as usize + 1]);
@@ -83,24 +113,47 @@ where
     }
 
     fn weekday(&mut self) -> Result<u8, Self::Error> {
+        let data = self.iface.read_register(Register::MONTH)?;
+
+        let century = data & BitFlags::CENTURY;
+        if century != 0 {
+            return Err(Error::InvalidDeviceCentury);
+        }
+
         self.read_register_decimal(Register::DOW)
     }
 
     fn day(&mut self) -> Result<u8, Self::Error> {
+        let data = self.iface.read_register(Register::MONTH)?;
+
+        let century = data & BitFlags::CENTURY;
+        if century != 0 {
+            return Err(Error::InvalidDeviceCentury);
+        }
+
         self.read_register_decimal(Register::DOM)
     }
 
     fn month(&mut self) -> Result<u8, Self::Error> {
         let data = self.iface.read_register(Register::MONTH)?;
-        let value = data & !BitFlags::CENTURY;
-        Ok(packed_bcd_to_decimal(value))
+
+        let century = data & BitFlags::CENTURY;
+        if century != 0 {
+            return Err(Error::InvalidDeviceCentury);
+        }
+
+        Ok(packed_bcd_to_decimal(data))
     }
 
     fn year(&mut self) -> Result<u16, Self::Error> {
-        let mut data = [0; 3];
-        data[0] = Register::MONTH;
-        self.iface.read_data(&mut data)?;
-        Ok(year_from_registers(data[1], data[2]))
+        let data = self.iface.read_register(Register::MONTH)?;
+
+        let century = data & BitFlags::CENTURY;
+        if century != 0 {
+            return Err(Error::InvalidDeviceCentury);
+        }
+
+        Ok(2000 + (self.read_register_decimal(Register::YEAR)? as u16))
     }
 
     fn date(&mut self) -> Result<NaiveDate, Self::Error> {
@@ -109,12 +162,15 @@ where
         self.iface.read_data(&mut data)?;
 
         let offset = Register::DOM as usize;
-        let year = year_from_registers(
-            data[Register::MONTH as usize + 1 - offset],
-            data[Register::YEAR as usize + 1 - offset],
-        );
-        let month =
-            packed_bcd_to_decimal(data[Register::MONTH as usize + 1 - offset] & !BitFlags::CENTURY);
+
+        let century = data[Register::MONTH as usize + 1 - offset] & BitFlags::CENTURY;
+        if century != 0 {
+            return Err(Error::InvalidDeviceCentury);
+        }
+
+        let year =
+            2000 + (packed_bcd_to_decimal(data[Register::YEAR as usize + 1 - offset]) as u16);
+        let month = packed_bcd_to_decimal(data[Register::MONTH as usize + 1 - offset]);
         let day = packed_bcd_to_decimal(data[Register::DOM as usize + 1 - offset]);
         let date = NaiveDate::from_ymd_opt(year.into(), month.into(), day.into());
         some_or_invalid_error(date)
@@ -174,39 +230,30 @@ where
     }
 
     fn set_year(&mut self, year: u16) -> Result<(), Self::Error> {
-        if !(2000..=2100).contains(&year) {
+        if !(2000..=2099).contains(&year) {
             return Err(Error::InvalidInputData);
         }
         let data = self.iface.read_register(Register::MONTH)?;
         let month_bcd = data & !BitFlags::CENTURY;
-        if year > 2099 {
-            let mut data = [
-                Register::MONTH,
-                BitFlags::CENTURY | month_bcd,
-                decimal_to_packed_bcd((year - 2100) as u8),
-            ];
-            self.iface.write_data(&mut data)
-        } else {
-            let mut data = [
-                Register::MONTH,
-                month_bcd,
-                decimal_to_packed_bcd((year - 2000) as u8),
-            ];
-            self.iface.write_data(&mut data)
-        }
+
+        let mut data = [
+            Register::MONTH,
+            month_bcd,
+            decimal_to_packed_bcd((year - 2000) as u8),
+        ];
+        self.iface.write_data(&mut data)
     }
 
     fn set_date(&mut self, date: &rtcc::NaiveDate) -> Result<(), Self::Error> {
-        if date.year() < 2000 || date.year() > 2100 {
+        if !(2000..=2099).contains(&date.year()) {
             return Err(Error::InvalidInputData);
         }
-        let (month, year) = month_year_to_registers(date.month() as u8, date.year() as u16);
         let mut payload = [
             Register::DOW,
             date.weekday().number_from_sunday() as u8,
             decimal_to_packed_bcd(date.day() as u8),
-            month,
-            year,
+            decimal_to_packed_bcd(date.month() as u8),
+            decimal_to_packed_bcd((date.year() - 2000) as u8),
         ];
         self.iface.write_data(&mut payload)
     }
@@ -238,28 +285,6 @@ fn hours_from_register(data: u8) -> Hours {
         Hours::PM(packed_bcd_to_decimal(
             data & !(BitFlags::H24_H12 | BitFlags::AM_PM),
         ))
-    }
-}
-
-fn year_from_registers(month: u8, year: u8) -> u16 {
-    let century = month & BitFlags::CENTURY;
-    let year = packed_bcd_to_decimal(year);
-    if century != 0 {
-        2100 + u16::from(year)
-    } else {
-        2000 + u16::from(year)
-    }
-}
-
-fn month_year_to_registers(month: u8, year: u16) -> (u8, u8) {
-    if year > 2099 {
-        let month = BitFlags::CENTURY | decimal_to_packed_bcd(month);
-        (month, decimal_to_packed_bcd((year - 2100) as u8))
-    } else {
-        (
-            decimal_to_packed_bcd(month),
-            decimal_to_packed_bcd((year - 2000) as u8),
-        )
     }
 }